+++ src/Core/Defaults.cs
@@ -7,7 +7,7 @@ namespace GreenDonut
         public static readonly int CacheSize = 1000;
         public static readonly TimeSpan BatchRequestDelay = TimeSpan
             .FromMilliseconds(50);
-        public static readonly int MinimumCacheSize = 10;
+        public static readonly int MinimumCacheSize = 1;
         public static readonly TimeSpan SlidingExpiration = TimeSpan.Zero;
     }
 }